feat: move setup vault page to common package
