--- clients/extension/src/pages/popup/pages/setup-vault/type/state/vaultSecurityType.ts
@@ -1,19 +1,0 @@
-import { useCorePathParams } from '@core/ui/navigation/hooks/useCorePathParams'
-import {
-  defaultVaultSecurityType,
-  VaultSecurityType,
-} from '@core/ui/vault/VaultSecurityType'
-import { useCallback } from 'react'
-
-export const useVaultSecurityType = () => {
-  const [{ type }, setParams] = useCorePathParams<'setupVault'>()
-
-  const value: VaultSecurityType = type ?? defaultVaultSecurityType
-
-  const setValue = useCallback(
-    (type: VaultSecurityType) => setParams(prev => ({ ...prev, type })),
-    [setParams]
-  )
-
-  return [value, setValue] as const
-}
