+++ core/ui/vault/create/setup-vault/hooks/useSetupVaultPageAnimation.ts
@@ -1,11 +1,15 @@
 import { useRive, useStateMachineInput } from '@rive-app/react-canvas'
-import { useEffect, useState } from 'react'
+import { useCallback, useEffect, useState } from 'react'
 
 const STATE_MACHINE_NAME = 'State Machine 1'
 const STATE_INPUT_NAME = 'Switch'
 
+// @antonio: required by product to switch animation initially. If switched too fast, the animation breaks we need a delay.
+const INITIAL_SWITCH_DELAY_MS = 360
+
 export const useSetupVaultPageAnimation = () => {
   const [isPlaying, setIsPlaying] = useState(false)
+  const [isReady, setIsReady] = useState(false)
 
   const { RiveComponent, rive } = useRive({
     src: '/assets/animations/choose-vault/index.riv',
@@ -19,15 +23,31 @@ export const useSetupVaultPageAnimation = () => {
     STATE_INPUT_NAME
   )
 
-  const onPlay = () => setIsPlaying(true)
+  useEffect(() => {
+    if (!stateMachineInput) return
+
+    setTimeout(() => {
+      stateMachineInput?.fire()
+      setIsReady(true)
+    }, INITIAL_SWITCH_DELAY_MS)
+  }, [stateMachineInput])
+
+  const onPlay = useCallback(() => {
+    setIsReady(true)
+    stateMachineInput?.fire()
+    setIsPlaying(true)
+  }, [stateMachineInput])
 
   useEffect(() => {
     if (!isPlaying) return
-
     const timeoutId = setTimeout(() => setIsPlaying(false), 1400)
-
     return () => clearTimeout(timeoutId)
   }, [isPlaying])
 
-  return { RiveComponent, stateMachineInput, isPlaying, onPlay }
+  return {
+    isReady,
+    RiveComponent: RiveComponent,
+    isPlaying,
+    onPlay,
+  }
 }
